fix: card image
